requirements.txt
albumentations==1.4.22
clip==0.2.0
comet_ml==3.47.5
coremltools==8.1
duckdb==1.1.3
dvclive==3.48.0
flatbuffers==23.5.26
hub_sdk==0.0.17
ipython==8.14.0
ipython==8.12.3
lancedb==0.17.0
lap==0.5.12
lvis==0.5.3
matplotlib==3.10.0
mss==10.0.0
ncnn==1.0.20240820
nncf==2.14.0
numpy==2.2.0
onnx==1.17.0
onnx2tf==1.26.3
onnxruntime==1.20.1
onnxslim==0.1.44
openai==0.27.8
opencv_python==4.9.0.80
openvino==2024.5.0
paddle==1.0.2
pafy==0.5.5
pandas==1.5.3
Pillow==11.0.0
psutil==5.9.0
psutil==5.9.8
py_cpuinfo==9.0.0
pycocotools==2.0.8
pytubefix==8.8.1
PyYAML==6.0.1
PyYAML==6.0.2
ray==2.10.0
Requests==2.32.3
roboflow==1.1.50
sahi==0.11.18
scipy==1.14.1
seaborn==0.13.2
sentry_sdk==2.19.2
Shapely==2.0.6
streamlit==1.41.1
streamlit_select==0.3
super_gradients==3.7.1
tensorflow==2.8.0
tensorflowjs==4.22.0
tensorrt==10.7.0
tflite_runtime==2.14.0
tflite_support==0.4.4
torch==2.2.1
torchvision==0.17.1
tqdm==4.67.1
tritonclient==2.52.0
ultralytics_thop==2.0.8
wandb==0.19.1
x2paddle==1.6.0
yt_dlp==2024.12.13
